Make freq chunking the default
